Fix a bug with reftex where tex-info and LaTeX interferred.
* tex-info.el (Texinfo-reftex-hook): Replace use of
reftex-tables-dirty by that of reftex-default-label-alist-entries
to trigger call to reftex-compile-variables in a standard way.
The problem was that generating a TOC for a Texinfo file and then
for a LaTeX file was not working for the LaTeX file because the
Texinfo file was using LaTeX label style and as such
reftex-ensure-compiled-variables was not calling
reftex-compile-variables, which caused wrong
reftex-everything-regexp for sections.

Signed-off-by: Tassilo Horn <tsdh@gnu.org>