Show and hide warnings and bad boxes in error overview
* tex-buf.el (TeX-error-list-skip-warning-p): New function.
(TeX-parse-TeX): Skip warnings or bad boxes if they are to be ignored.
(TeX-error-list): Describe information present in the list.
(TeX-parse-error): When storing warnings and bad boxes, add them to
`TeX-error-list' even if `TeX-debug-bad-boxes' and `TeX-debug-warnings'
are nil.
(TeX-error-overview-make-entries): Skip warnings or bad boxes if they
are to be ignored.
(TeX-error-overview-toggle-debug-bad-boxes):
(TeX-error-overview-toggle-debug-warnings): New functions.
(TeX-error-overview-mode-map, TeX-error-overview-menu): Add
`TeX-error-overview-toggle-debug-bad-boxes' and
`TeX-error-overview-toggle-debug-warnings'.
(TeX-error-overview): Make the list of entries for the overview before
creating the buffer and use this list to decide whether there is
something to show.