Improve caption and label insertion in style/caption.el
* style/caption.el (LaTeX-caption-key-val-options): Add key slc
which is an abbrev for singlelinecheck.
(LaTeX-arg-caption-captionbox): Delete unused prompt argument.
Ask for short caption acc. to the value of
`LaTeX-short-caption-prompt-length'.  Use `LaTeX-label' instead of
`TeX-insert-macro' to insert a label.  Use `TeX-argument-insert'
instead of conditional insert.
("caption"): Do not fontify the last argument of \captionbox.
Fontification happens within \begin...\end{tabular} or
\includegraphics.