Add `NO-INSERT' argument to `LaTeX-label'
* doc/changes.texi: Document changes to `LaTeX-label' and
`LaTeX-label-function'.

* latex.el (LaTeX-label): Add second optional argument NO-INSERT.
This argument is passed to function bound to
`LaTeX-label-function' (usually `reftex-label' with RefTeX
enabled).  If `LaTeX-label-function' is nil and NO-INSERT is
non-nil, label is read and returned as a string.  Remove some
unnecessary checks.
(LaTeX-label-function): Mention in docstring that the function is
expected to take an optional argument NO-INSERT.  This is an
incompatible change.