Fix auto selection of japanese TeX engine
* tex-jp.el (japanese-latex-mode-initialization): Use
`LaTeX-match-class-option' to detect uplatex option given to
jsarticle, jsreport and jsbook class.  (The root of this problem is
the same as Bug#25563)
(): Quit appending to `LaTeX-command-style' because this made the
user's intentional selection of `TeX-engine' ineffective with regard
to the selection of the executable names.
Fix copyright year.
; Add Ikumi Keita to maintainer of tex-jp.el.
* doc/auctex.texi (Japanese): Add description of Japanese TeX engine:
How to use, change and set the value of the engine.
Add description of the customize options related to the encoding of
Japanese TeX documents.