Unify treatment of face argument over the relevant functions.
* font-latex.el (font-latex-keyword-matcher): Conform the face
argument correctly to the format `font-lock-keywords' requires.
(font-latex-make-user-keywords) : Call `font-latex-make-match-defun'
directly without `eval'.
(font-latex-deactivated-keyword-classes) : Accept a face symbol, as
well as a form returning a face symbol, for the third element for each
item of `font-latex-built-in-keyword-classes'.
(font-latex-built-in-keyword-classes) : Remove quoting from the face
names.  Contrary to the sentence of the doc string
"The third element is the symbol of a face to be used or a Lisp
form returning a face symbol.", symbol of a face was invalid due to
the above bugs.