Don't alter standard value or saved value themselves of customize option
* tex-jp.el (): Stop using `setcar' on the elements of
`TeX-command-list'.  This used to alter the standard value
or saved value themselves of the customize option since
`setcar' operates descructively on lists as Lisp objects.
Instead create new list by `mapcar' and `append'.
(): Deal with the case that `TeX-view-program-list-builtin'
doesn't have an entry of "xdvi".
Change the expected executable name to "%(xdvi)" as well.