Resolve preview-latex incompatibility with Japanese TeX (Bug#25322)
* preview.el.in (preview-error-quote): Work consistently with
decoded text.  The idea of "encoding entire string beforehand and
decoding it at the last stage" was used previously, but that failed
with `japanese-shift-jis' coding system containing bytes which
happen to coincide with regexp meta characters.
(preview--decode-^^ab):
(preview--convert-^^ab): New functions.
(TeX-inline-preview-internal): Record the process coding system
for decode assigned already so that `preview-error-quote' can
decode the given text correctly afterward.  To achive that, change
the role of `preview-coding-system' together with the changes to
`preview-error-quote'.
Give back the command options provided in `(TeX-engine-alist)' to
latex command when preamble caching is enabled.
(preview-coding-system): Modify doc string according to the
changes above to `preview-error-quote' and
`TeX-inline-preview-internal'.
* tex-buf.el (TeX-adjust-process-coding-system): New function.
Adjust the process coding system for asynchronous process launched
within AUCTeX.
(TeX-after-start-process-function): Change default value to
`TeX-adjust-process-coding-system'.
* tex-jp.el (japanese-TeX-set-process-coding-system): Specify end
of line format on the coding systems assigned to process.
(): Don't set `japanese-TeX-set-process-coding-system' to
`TeX-after-start-process-function'.  It is now called within
`TeX-adjust-process-coding-system'.
* prv-xemacs.el (preview-buffer-recoding-alist): Fix docstring.
(coding-system-change-eol-conversion): Add advice because XEmacs
21.4 mule-ucs fails to define `utf-8' coding system to respond
properly to this function.
(preview-ps-quote-filename): Add advice so that this function is
not confused by backslashes as path separator in the return value
of XEmacs function.
* tests/japanese/preview-latex.el:
* tests/japanese/preview-error-test.tex:
* tests/japanese/preview-error-test2.tex:
* tests/japanese/prv-dif-code.tex: New files for regression test.