Resolve conflict of argument spec between regexp and literal
* tex.el (TeX-search-files-kpathsea): Treat EXTENSIONS argument as a
list of regexps in accordance with other parts of AUCTeX.
(TeX-ispell-document): Treat `TeX-file-extensions' as a list of
regexps in accordance with other parts of AUCTeX.
(TeX-file-extensions, TeX-Biber-file-extensions):
(BibTeX-file-extensions, BibLaTeX-style-extensions):
(BibTeX-style-extensions): Change custom type to regexp from string.
(TeX-doc-extensions): Turn into list of regexps.
* latex.el (LaTeX-split-bibs): Treat `TeX-Biber-file-extensions' as a
list of regexps in accordance with other parts of AUCTeX.