Accept non-ascii file name in accord with change in TL 2018
* tex.el (TeX-expand-list-builtin): Add new entry %T.  Same as %t,
except to enclose with \detokenize{} for non UTF-8 LaTeX when \input
is supplmented.
Adjust the entries %` and %' so that \input is supplemented only when
any TeX code is present between them and leave the bind to
`TeX-command-text' for later examination.
The bind to `TeX-command-pos' is no longer retained.
(TeX-command-list): Use %T for "LaTeX".
Adjust "TeX" and "AmSTeX" as the same with "LaTeX" in the aspect that
user can supply one's own TeX code, as well as any command line
options, through `TeX-command-extra-options'.
* tex-buf.el (TeX--master-or-region-file-with-extra-quotes): New
function to act as a wrapper of `TeX-master-file' and
`TeX-region-file' inside `TeX-command-expand'.
(TeX-command-expand): Use the above function as the value of `file'
and get rid of tricky temporal overriding of `file' with lambda form.
(TeX-region-create): Make the first line parsing of %&FORMAT
construct, if any, to be valid even for region compilation.
Discard text properties when constructing the content of _region_.tex.
Drop bind check for `buffer-file-coding-system'.
* tests/tex/command-expansion.el (TeX-command-expansion): Reflect the
change that \input is not necessarily supplemented now by %`-%' pair
in `TeX-command-expand'.