Adjust preview-latex in case sub file isn't in master directory
* preview.el.in (preview-dvips-abort):
(preview-gs-dvips-sentinel):
(preview-dvipng-place-all): Make sure directory part of dvi file to be
deleted is removed because deletion takes place in master directory.
(preview-start-pdf2dsc): Make sure directory part of pdf file is
removed so that path for later use is constructed correctly.
(preview-cache-preamble):
(preview-region):
(preview-document): Retain directory part of the file name to be
passed to `preview-generate-preview' because `TeX-run-command' in
`TeX-inline-preview-internal' doesn't recognize the right path
otherwise.
(preview-generate-preview): Adjust in accord with the above change.