Allow the user to customize which TeX commands are available in each mode.
Commands such as Biber, DVI2PS, PS2PDF, Index, etc. are not useful in
ConTeXt.  ConTeXt generates a PDF directly and has its own indexing
mechanism.

In Mark IV (which has been the preferred version for over a decade and
is actively developed), a new bibliography package is also available,
which can process old bibtex files as well.  So there is no need for
BibTeX.

Having so many commands in context-mode clutters the menus.  In the
definition of TeX-command-list, we either provide a list of relevant
modes, or set the value to `t' indicating that this command should
always be available.  In the latter case, the user cannot customize
it.  Here we provide sensible defaults as well as the ability for the
user to customize.

* tex.el (TeX-command-list): Hide by default some commands from context-mode.

Signed-off-by: Mosè Giordano <mose@gnu.org>