Add check for empty overlay
* tex-fold.el (TeX-fold-hide-item): Add an extra check to make
sure that the function runs when overlay start and end are
non-nil.  What is done in the end of function isn't useful with
evaporated overlays and used to cause bugs because `overlay-start'
and `overlay-end' return nil for such overlays, and functions such
as `TeX-fold-overfull-p` try to move point to the
overlay-end. (bug#37337)