Improve conformance to code conventions, and documents
* font-latex.el (font-latex-update-math-env): Rename by changing "--"
to "-" in function name.
Don't alter user customize option `font-latex-math-environments'. In
order to achieve that, use always `texmathp-tex-commands1' for
input and omit argument LIST.
* style/empheq.el: Arrange in accord with the above changes.
* style/amsmath.el:
* style/breqn.el:
* style/mathtools.el:
Arrange in accord with the above changes.
Use `cl-pushnew' instead of `add-to-list' and require cl-lib in order
to do that.
* doc/changes.texi (News in 12.3):
* texmathp.el:
Mention change about support for AMS-LaTeX and packages mathtools,
empheq and breqn.