Discard `TeX-auto-update' to silence byte compiler
Now that we add `TeX-safe-auto-write' as local hook in
`write-contents-functions', we don't have to check every time whether
it is AUCTeX buffer.

* latex.el (BibTeX-auto-store):
* tex-info.el (TeX-texinfo-mode):
Don't set `TeX-auto-update'.
* tex.el (VirTeX-common-initialization): Don't set `TeX-auto-update'.
(TeX-safe-auto-write): Don't check whether it is AUCTeX buffer or
BibTeX buffer.
(TeX-update-style): Consult `major-mode' instead of `TeX-auto-update'
to check whether it is BibTeX buffer.
* bib-cite.el: Don't use `TeX-auto-update'. Consult `TeX-mode-p' to
check whether it is AUCTeX buffer.