* font-latex.el (font-latex-make-built-in-keywords): Make it a macro
This lets us expose the code it generates to the compiler without
having to call `byte-compile` explicitly.  It also reduces the reliance
on `eval`, saving the souls of many kitten along the way.

(font-latex-built-in-keyword-classes): Define it at compile-time as well.
(font-latex-keywords-1, font-latex-keywords-2):
Move their definition to `font-latex-make-built-in-keywords`.
(font-latex-keyword-matcher): Define it at compile-time as well.
(font-latex--make-match-defun): Rename from `font-latex-make-match-defun`.
Define it at compile-time as well.  Return the function definition (as
the docstring suggests) rather than evaluating it.
(font-latex-make-user-keywords): Adjust call to it.
(<toplevel>): Don't manually call the byte-compiler.