Remove some XEmacs and old-Emacs code
Remove uses of `easy-menu-add` and `easy-menu-remove` which have
always been no-op in Emacs and are declared obsolete in Emacs-28.

* bib-cite.el (bib-cite-setup-highlight-mouse-keymap): Remove Emacs-19 code.

* tex-buf.el (TeX-region-create): Remove use of variable
`font-lock-maximum-size` declared obsolete in Emacs-24.

* toolbar-x.el: Remove left over code from XEmacs compatibility.
Remove XEmacs-only `:toolbar` and `:dropdown-toolbar` as well
`:insert` which was only used to select Emacs-only or XEmacs-only entries.
(toolbarx-default-toolbar-meaning-alist): Replace all uses of
`[Emacs-only XEmacs-only]` vectors with the `Emacs-only` form.