Use DEFAULT argument in latex.el where appropriate
* latex.el (LaTeX-env-array, LaTeX-env-tabular*): Use the DEFAULT
argument of `TeX-read-string' when asking for the mandatory FMT
argument.
(LaTeX-env-minipage): Use `format' instead of `concat' to assemble
the prompt.
(LaTeX-env-bib): Use DEFAULT argument for the bibitem label.
(TeX-arg-file-name, TeX-arg-file-name-sans-extension): Use the
DEFAULT argument only when used as a mandatory argument.
(TeX-arg-date, TeX-arg-version): Add support for DEFAULT argument
of `TeX-read-string'.