Don't use void function texinfo-outline-level
Apparently, the function `texinfo-outline-level' was added to
texinfo.el in 1996 (ChangeLog.6) and removed in 2003 (ChangeLog.10).
In release 9.4f (1996) `outline-level' was set to this then newly
added function, which has been void for 19 years now.

I just noticed it now, because consult [0] checks if `outline-level' is
bound, in which case it is funcall'ed.  In .texi buffer with auctex
loaded, this triggers an error.

[0] https://elpa.gnu.org/packages/consult.html

* tex-info.el (TeX-texinfo-mode): Don't bind texinfo-outline-level to
outline-level.