Simplify implementation of style/enumitem.el
* style/enumitem.el (LaTeX-enumitem-env-with-opts): Delete
function.
(LaTeX-arg-enumitem-SetEnumitemKey): Query and insert only the
first mandatory argument.  The second on is queried in the hook.
(LaTeX-arg-enumitem-SetEnumitemValue): Query and insert only the
second mandatory argument.  The others are queried in the hook.
(LaTeX-arg-enumitem-setlist): Delete function, now performed in
the hook.
(LaTeX-enumitem-auto-cleanup):
("enumitem"): Use the function `LaTeX-env-item-args' instead of
`LaTeX-enumitem-env-with-opts'.