Fix `TeX-documentation-texdoc' for okular
We used `start-process-shell-command' for these 4 years.  However,
it turned out that that invalidated okular.  (See bug#40577 and
<URL:https://lists.gnu.org/r/auctex/2023-01/msg00006.html>.)

The reason that `TeX-documentation-texdoc' tried hard to collect and
show the output from Texdoc is that the exit code wasn't meaningful
once[1].  This was the only way to notify the user when Texdoc can't
find any documentation for the given keyword.

However, recent Texdoc is improved to return non-zero exit code when
it can't find any documentation[2].  Now we use `call-process' instead
of `start-process-shell-command'.  This has a drawback that user who
sticks to older TeX Live distribution isn't notified at all when the
given keyword didn't match any documentation, but we hope that it
doesn't cause major problem.

[1] https://debbugs.gnu.org/cgi/bugreport.cgi?bug=28905#17
[2] https://tug.org/texdoc/doc/texdoc.man1.pdf

* tex.el (TeX-documentation-texdoc): Use `call-process' to invoke
Texdoc.
Move `executable-find' test up in the function body so that it covers
no prefix case as well.
Add NO-PROPERTIES argument to `thing-at-point'.
Remove unnecessary `save-excursion'.
Refill.