Arrange indent of tabular-like environments (bug#65648)
This fixes bug#54949 as well.

* latex.el (LaTeX-env-beginning-pos-indent): New function to align "&"
and \end{foo} relative to the indent of the line where \begin{foo} is,
not the column of \begin{foo} itself.
(LaTeX-indent-tabular, LaTeX-hanging-ampersand-position): Use the new
function instead of `LaTeX-env-beginning-pos-col'.  This makes
difference when \begin{foo} isn't on its own dedicated line.
(LaTeX-hanging-ampersand-position): Arrange the doc string.
* tests/latex/tabular-out.tex: Modify the expected result according to
the above change.