Throw out fold spec after a single failure
So far, TeX-fold-expand-spec did not abort its expansion of the
current spec if *any* of the expansions succeeded.  This can lead to
some unexpected behaviour.  For example, given

  (setopt TeX-fold-math-spec-list
          '(("[1] of {1}||{1}" ("cite"))))

expanding `\cite{this}' would result in the overlay displaying
`… of this' instead of just `this', because one of the arguments
(the mandatory one) succeeded.  This is probably not what the user
intended.  Instead, in order for the spec to be chosen, *all
arguments should succeed.

* tex-fold.el (TeX-fold-expand-spec): Return non-nil value only when
all expansions suceeded.