Amend TeX electric commands to allow alternative electric modes
1. In particular, we are splitting out the electric logic of
`TeX-insert-dollar'.

2. We've clarified that `TeX-refuse-unmatched-dollar' doesn't have an
effect when `TeX-electric-math' is non-nil.  The reason is that it's
complicated to support all math delimiters consistently, and the need
for this feature is unclear.  If there is a demand to support this in
the future, it can be done.  It's explained below in Appendix I.

3. We've removed the special case where
> If [[`TeX-electric-math']] non-nil and point is inside math
> mode right between a couple of single dollars, pressing `$'
> will insert another pair of dollar signs and leave the point
> between them.

4. We've changed `C-u $' to insert four dollars as opposed to one.  This
behavior is in-line with other symbols, in particular
`TeX-insert-quote' and `TeX-insert-backslash'.  If someone needs one
literal `$', they can type `C-1 $' or `C-q $'.

5. We've set appropriate `delete-selection' properties for
`TeX-insert-backslash', `TeX-newline' and `TeX-insert-dollar'.  See the
backstory below or the docstring for
`TeX--put-electric-delete-selection'.

6. We didn't fix `TeX-insert-quote'. Maybe later.

The backstory.

When a function bound to SYMBOL has optional «electric»
behavior, it might interfere with other «electric» modes,
e.g. `electric-pair-mode', `smartparens-mode'; see bug#47936.

As a way to «override» those modes, we use raw `insert' instead
of `self-insert-command'.  That prevents those electric modes
from running their hooks tied to `self-insert-command'.

However, when /our/ electric behavior is disabled (ELECTRICP is
nil), we want other electric modes to operate freely.  That
means, on the non-electric path, we should use
`self-insert-command' instead of `insert'.

Now, there arises an issue of `delete-selection'.  The electric
path usually doesn't want to delete selection, it wants to
operate some electricity on it; see bug#36385, bug#23177.  Now,
we could think that `delete-selection' for the non-electric path
should be t.  That would disable other electric modes from
working, as they also need to operate on selection.  The decision
is to inherit `delete-selection' from `self-insert-command',
which queries hooks from other electric modes to determine
whether deletion is necessary.

Appendix I.

To support a (generalized) TeX-insert-dollar with electric-math, we
would want the following.  We would want to forbid /arbitrary/ math mode
nesting, independent of the delimiter in TeX-electric-math.  However,
the naive approach causes issues with display math via "$$ bla-bla $$".
When the user presses $ once, they get "$-!-$".  When they press $
again, they can't get "$$-!-$$", because that is considered as "nested
math".  A simple way to fix this would be to special-case handling of a
dollar insert.  However, a more satisfactory way would be to "cycle"
through inline math, display math and no math on $ presses.  We actually
already implement that behavior on $ presses on a region.

* tex.el: (TeX-insert-dollar-electric-region, TeX-insert-dollar-electric):
(TeX--blink-matching-dollar, TeX-insert-dollar-action):
New function to refactor `TeX-insert-dollar'.
(TeX-insert-dollar): Use them.
(TeX-insert-dollar-1): Remove.
(TeX-electric-math, TeX-refuse-unmatched-dollar): Adjust the doc stirng.
(): Add appropriate `delete-selection' property to
`TeX-insert-backslash', `TeX-insert-dollar' and `TeX-insert-quote'.