Improve in-buffer completion of arguments
* latex.el (LaTeX-completion-macro-delimiters): Use `pcase' which
is more concise.
(LaTeX-completion-candidates-key-val)
(LaTeX-completion-candidates-completing-read-multiple)
(LaTeX-completion-candidates-completing-read): Add the property
":exclusive no" to the returned lists.  This is necessary since
`LaTeX--arguments-completion-at-point' is now prepended to
`completion-at-point-functions' and if it fails,
`TeX--completion-at-point' will still take over.
(LaTeX-completion-candidates-completing-read-multiple)
(LaTeX-completion-candidates-completing-read): Recognize backslash
as boundary which is necessary due to a change of behavior in
Emacs.  See discussion in this thread:
https://lists.gnu.org/archive/html/auctex-devel/2025-10/msg00013.html
Many thanks to Stefan Monnier <monnier@iro.umontreal.ca> for his
valuable comments.
(LaTeX-common-initialization): Prepend
`LaTeX--arguments-completion-at-point' to
`completion-at-point-functions'.